fix bug , change snackbar to dialogbox

+++ lib/payment/screens/payment_screen.dart
@@ -51,24 +51,27 @@ class PaymentScreen extends StatelessWidget {
                   icon: const Icon(Icons.edit)),
               IconButton(
                   onPressed: () {
-                    Get.defaultDialog(
-                        title: "Delete!",
-                        middleText: "Are you sure to delete!",
-                        actions: [
-                          TextButton(
-                            onPressed: () {
-                              Get.back();
-                            },
-                            child: const Text("Cancel"),
-                          ),
-                          TextButton(
-                            onPressed: () {
-                              paymentController.deletePayment(payment.id);
-                              Get.back();
-                            },
-                            child: const Text("Ok"),
-                          ),
-                        ]);
+                    Get.dialog(
+                      AlertDialog(
+                          title: const Text("Delete!"),
+                          content: const Text("Are you sure to delete!"),
+                          actions: [
+                            TextButton(
+                              onPressed: () {
+                                Get.back();
+                              },
+                              child: const Text("Cancel"),
+                            ),
+                            TextButton(
+                              onPressed: () {
+                                paymentController.deletePayment(payment.id);
+                                Get.back();
+                              },
+                              child: const Text("Ok"),
+                            ),
+                          ]
+                      )
+                    );
                   },
                   icon: const Icon(Icons.delete)),
             ],
